Core re-write for various reasons
While the intial version of Skills that was deployed to Almura on 9/7 was mostly stable, the code was done in a rush. I was not happy with how it looked and function so I set out to fix that. The new codebase, from my testing, is far more rigid and structured. The API has been expanded and the plugin implementation makes full use of it.

This is the first big step on the road to turning Skills into a platform.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>

Refactor configuration, add save-interval and documentation.

Fix commands not querying for Event.class for messages/effects.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>