Hot patch multi-click crafting exploit.
The issue is in Vanilla or perhaps where Sponge fires the CraftingEvent.Post.

This Mixin: https://github.com/SpongePowered/SpongeCommon/blob/stable-7/src/main/java/org/spongepowered/common/mixin/core/item/inventory/MixinSlotCrafting.java#L171

Basically this is being called when shift-click or left-click crafting. The problem is Vanilla passes the current stack of the cursor to this method which will keep going up and up in quantity and make the result stack passed to the crafting event keep going up and up. Shift clicking is not affected as a stack does not remain on the cursor.

This commit uses what the recipe says will be the result, ignoring the result stack until Sponge is fixed.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>