Bandaid fix for Vanilla attachables.
It is not trivial for the Sponge PhaseTracker to give a meaningful Cause
when dealing with physics as that is not the same tick. Since it can't,
someone could easily break a block underneath a crop they do not have
the level for and basically get infinite drops.

As a bandaid, if we're breaking a block we now check the attached block
on the TOP face of the cube and process that snapshot if and only if the
transaction wasn't being cancelled. If that processing returns a
CANCELLED state that means if they had broke the crop, it would have
been stopped. So we stop the anchor block from being broken we prevents
the entirety of the chain from occurring.

This isn't a permanent fix, I will state, as it does not take into
consideration the other faces that may have an attachable.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>